requirements.txt
kivymd
easygui
deepdiff